weniger code im unload der Webseite
git-svn-id: https://dev.plt.rwth-aachen.de/acplt-repo@5845 7a3915f3-eddd-4ae9-ab37-eb2ea51f0265

+++ history/develop/hmi/hmiJavaScript/trunk/JavaScript/hmi-class-HMI.js
@@ -1584,17 +1584,9 @@ HMI.prototype = {
 		unload
 	*********************************/
 	unload: function () {
-		this.hmi_log_trace("HMI.prototype.unload - Start");
-		
-		if (HMI){
-			HMI.PossServers = null;
-			HMI.PossSheets = null;
-			
-			if (HMI.KSClient){
-				HMI.KSClient.destroy();
-			}
+		if (HMI && HMI.KSClient){
+			HMI.KSClient.destroy();
 		}
-		this.hmi_log_trace("HMI.prototype.unload - End");
 	},
 	
 	/********************************************************************
